+++ e2e/runner/evm.go
@@ -98,7 +98,8 @@ func (runner *E2ERunner) DepositERC20WithAmountAndMessage(to ethcommon.Address, 
 	runner.Logger.Info("ERC20 Approve receipt tx hash: %s", tx.Hash().Hex())
 
 	tx, err = runner.ERC20Custody.Deposit(runner.EVMAuth, to.Bytes(), runner.ERC20Addr, amount, msg)
-	runner.Logger.Print("TX: %v", tx)
+	runner.Logger.Info("TX: %v", tx)
+
 	if err != nil {
 		panic(err)
 	}